# old-replit-workspace/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Common Ground Community Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events
      parameters:
        - { name: query, in: query, schema: { type: string } }
        - { name: city, in: query, schema: { type: string } }
        - { name: county, in: query, schema: { type: string } }
        - { name: category, in: query, schema: { type: string } }
        - { name: free, in: query, schema: { type: string } }
        - { name: kids, in: query, schema: { type: string } }
        - { name: outdoor, in: query, schema: { type: string } }
        - { name: from, in: query, schema: { type: string } }
        - { name: to, in: query, schema: { type: string } }
        - { name: sort, in: query, schema: { type: string } }
        - { name: month, in: query, schema: { type: string } }
        - { name: view, in: query, schema: { type: string } }
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event by ID
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Event detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /activities:
    get:
      operationId: listActivities
      tags: [events]
      summary: List activity events
      parameters:
        - { name: query, in: query, schema: { type: string } }
        - { name: city, in: query, schema: { type: string } }
        - { name: county, in: query, schema: { type: string } }
        - { name: category, in: query, schema: { type: string } }
        - { name: free, in: query, schema: { type: string } }
        - { name: kids, in: query, schema: { type: string } }
        - { name: outdoor, in: query, schema: { type: string } }
        - { name: from, in: query, schema: { type: string } }
        - { name: to, in: query, schema: { type: string } }
        - { name: sort, in: query, schema: { type: string } }
        - { name: month, in: query, schema: { type: string } }
      responses:
        "200":
          description: List of activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
  /calendar-events:
    get:
      operationId: listCalendarEvents
      tags: [events]
      summary: List events for calendar view
      parameters:
        - { name: month, in: query, schema: { type: string } }
        - { name: activity, in: query, schema: { type: string } }
        - { name: city, in: query, schema: { type: string } }
        - { name: county, in: query, schema: { type: string } }
        - { name: category, in: query, schema: { type: string } }
      responses:
        "200":
          description: Events for calendar
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List alerts
      parameters:
        - { name: city, in: query, schema: { type: string } }
        - { name: county, in: query, schema: { type: string } }
        - { name: type, in: query, schema: { type: string } }
      responses:
        "200":
          description: Alerts grouped by status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertsResult"
  /meetings:
    get:
      operationId: listMeetings
      tags: [meetings]
      summary: List meetings
      parameters:
        - { name: city, in: query, schema: { type: string } }
        - { name: county, in: query, schema: { type: string } }
        - { name: body, in: query, schema: { type: string } }
        - { name: type, in: query, schema: { type: string } }
      responses:
        "200":
          description: Meetings grouped by status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingsResult"
  /meetings/{id}:
    get:
      operationId: getMeeting
      tags: [meetings]
      summary: Get meeting by ID
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Meeting detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /volunteer:
    get:
      operationId: listVolunteer
      tags: [volunteer]
      summary: List volunteer opportunities
      parameters:
        - { name: city, in: query, schema: { type: string } }
        - { name: county, in: query, schema: { type: string } }
      responses:
        "200":
          description: List of volunteer opportunities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VolunteerOpportunity"
...
  /search:
    get:
      operationId: globalSearch
      tags: [search]
      summary: Global search across all content
      parameters:
        - { name: query, in: query, schema: { type: string } }
        - { name: city, in: query, schema: { type: string } }
        - { name: county, in: query, schema: { type: string } }
        - { name: category, in: query, schema: { type: string } }
        - { name: free, in: query, schema: { type: string } }
        - { name: kids, in: query, schema: { type: string } }
        - { name: from, in: query, schema: { type: string } }
        - { name: to, in: query, schema: { type: string } }
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResults"
  /submit:
    post:
      operationId: submitItem
      tags: [submissions]
      summary: Submit a community item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitItemInput"
      responses:
        "200":
          description: Submission accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitResult"
  /subscribe:
    post:
      operationId: subscribe
      tags: [submissions]
      summary: Subscribe to the weekly digest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscribeInput"
      responses:
        "200":
          description: Subscription saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscribeResult"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Event:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: ["string", "null"] }
        category: { type: string }
        status: { type: string }
        startDateTime: { type: string, format: date-time }
        endDateTime: { type: ["string", "null"], format: date-time }
        locationName: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        county: { type: ["string", "null"] }
        cost: { type: ["string", "null"] }
        isFree: { type: boolean }
        isKidFriendly: { type: boolean }
        isOutdoor: { type: boolean }
        tags: { type: ["string", "null"] }
        sourceUrl: { type: ["string", "null"] }
        sourceName: { type: ["string", "null"] }
        originalUrl: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, title, category, status, startDateTime, isFree, isKidFriendly, isOutdoor, createdAt, updatedAt]
    Alert:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: ["string", "null"] }
        alertType: { type: string }
        severity: { type: string }
        status: { type: string }
        city: { type: ["string", "null"] }
        county: { type: ["string", "null"] }
        startsAt: { type: string, format: date-time }
        expiresAt: { type: ["string", "null"], format: date-time }
        sourceUrl: { type: ["string", "null"] }
        sourceName: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, title, alertType, severity, status, startsAt, createdAt, updatedAt]
    Meeting:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        governmentBody: { type: string }
        meetingType: { type: string }
        status: { type: string }
        startDateTime: { type: string, format: date-time }
        endDateTime: { type: ["string", "null"], format: date-time }
        locationName: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        county: { type: ["string", "null"] }
        agendaUrl: { type: ["string", "null"] }
        minutesUrl: { type: ["string", "null"] }
        videoUrl: { type: ["string", "null"] }
        originalUrl: { type: ["string", "null"] }
        sourceUrl: { type: ["string", "null"] }
        sourceName: { type: ["string", "null"] }
        summary: { type: ["string", "null"] }
        plainEnglishSummary: { type: ["string", "null"] }
        whyResidentsCare: { type: ["string", "null"] }
        keyTopics: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, title, governmentBody, meetingType, status, startDateTime, createdAt, updatedAt]
    VolunteerOpportunity:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        organization: { type: string }
        description: { type: ["string", "null"] }
        dateTime: { type: ["string", "null"], format: date-time }
        locationName: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        county: { type: ["string", "null"] }
        status: { type: string }
        contactName: { type: ["string", "null"] }
        contactEmail: { type: ["string", "null"] }
        sourceUrl: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, title, organization, status, createdAt, updatedAt]
...
    AlertsResult:
      type: object
      properties:
        activeAlerts:
          type: array
          items:
            $ref: "#/components/schemas/Alert"
        expiredAlerts:
          type: array
          items:
            $ref: "#/components/schemas/Alert"
      required: [activeAlerts, expiredAlerts]
    MeetingsResult:
      type: object
      properties:
        upcomingMeetings:
          type: array
          items:
            $ref: "#/components/schemas/Meeting"
        completedMeetings:
          type: array
          items:
            $ref: "#/components/schemas/Meeting"
        governmentBodies:
          type: array
          items:
            type: string
      required: [upcomingMeetings, completedMeetings, governmentBodies]
    SearchResults:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        alerts:
          type: array
          items:
            $ref: "#/components/schemas/Alert"
        meetings:
          type: array
          items:
            $ref: "#/components/schemas/Meeting"
        volunteer:
          type: array
          items:
            $ref: "#/components/schemas/VolunteerOpportunity"
      required: [events, alerts, meetings, volunteer]
    SubmitItemInput:
      type: object
      properties:
        submitterName: { type: string }
        submitterEmail: { type: string }
        submissionType: { type: string }
        category: { type: string }
        title: { type: string }
        description: { type: string }
        startDateTime: { type: string }
        endDateTime: { type: string }
        locationName: { type: string }
        address: { type: string }
        city: { type: string }
        county: { type: string }
        cost: { type: string }
        tags: { type: string }
        sourceUrl: { type: string }
        isFree: { type: boolean }
        isKidFriendly: { type: boolean }
        isOutdoor: { type: boolean }
      required: [submitterName, submitterEmail, title, startDateTime, city, county]
    SubmitResult:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
      required: [success, message]
    SubscribeInput:
      type: object
      properties:
        email: { type: string }
        city: { type: string }
        county: { type: string }
        interests: { type: string }
      required: [email]
    SubscribeResult:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
      required: [success, message]