# old-replit-workspace/lib/api-zod/src/generated/types/event.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Common Ground Community Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface Event {
  id: string;
  title: string;
  /** @nullable */
  description?: string | null;
  category: string;
  status: string;
  startDateTime: Date;
  /** @nullable */
  endDateTime?: Date | null;
  /** @nullable */
  locationName?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  county?: string | null;
  /** @nullable */
  cost?: string | null;
  isFree: boolean;
  isKidFriendly: boolean;
  isOutdoor: boolean;
  /** @nullable */
  tags?: string | null;
  /** @nullable */
  sourceUrl?: string | null;
  /** @nullable */
  sourceName?: string | null;
  /** @nullable */
  originalUrl?: string | null;
  createdAt: Date;
  updatedAt: Date;
}