# old-replit-workspace/lib/api-zod/src/generated/types/searchResults.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Common Ground Community Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { Alert } from './alert';
import type { Event } from './event';
import type { Meeting } from './meeting';
import type { VolunteerOpportunity } from './volunteerOpportunity';

export interface SearchResults {
  events: Event[];
  alerts: Alert[];
  meetings: Meeting[];
  volunteer: VolunteerOpportunity[];
}