# old-replit-workspace/lib/api-zod/src/generated/types/submitResult.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Common Ground Community Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface SubmitResult {
  success: boolean;
  message: string;
}