# old-replit-workspace/lib/api-zod/src/generated/types/subscribeInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Common Ground Community Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface SubscribeInput {
  email: string;
  city?: string;
  county?: string;
  interests?: string;
}